Cleanup s3ng options code
